# specs/001-core-system/contracts/host-api.openapi.yaml
openapi: 3.1.0
info:
  title: CameraCommander2 Host API
  version: 1.0.0
  description: |
    REST contract exposed by the headless Python host application. The web UI
    consumes this surface exclusively (Constitution Principle I). The CLI
    invokes the same underlying services so behaviour is identical.

    All hardware-affecting endpoints are serialised behind a single hardware
    worker thread. At most one job (timelapse or video pan) runs at a time
    (FR-039); a second launch returns HTTP 409.
servers:
  - url: http://localhost:8000
    description: Local Pi Zero 2 W host (default)
...
paths:
...
  /api/camera/settings:
    get:
      tags: [camera]
      summary: Query all current camera settings
      operationId: getCameraSettings
      responses:
        "200":
          description: Map of gphoto2 widget paths to current values, types, and choices.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraSettings"
        "503":
          $ref: "#/components/responses/CameraUnavailable"
    put:
      tags: [camera]
      summary: Apply a partial camera settings update
      operationId: putCameraSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CameraSettingsUpdate"
      responses:
        "200":
          description: Settings applied. The response echoes the resulting settings map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraSettings"
        "400":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/CameraUnavailable"
  /api/camera/capture:
    post:
      tags: [camera]
      summary: Trigger a one-shot still capture (test capture)
      operationId: postCameraCapture
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureRequest"
      responses:
        "200":
          description: Capture metadata; binary content fetched via `GET /api/camera/captures/{id}`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResult"
        "409":
          description: A running job currently owns the hardware lock.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/CameraUnavailable"
  /api/camera/captures/{capture_id}:
    get:
      tags: [camera]
      summary: Download a previously captured one-shot image
      operationId: getCaptureFile
      parameters:
        - in: path
          name: capture_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image bytes; content-type matches the camera output (image/jpeg or vendor RAW).
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/x-canon-cr3: { schema: { type: string, format: binary } }
            image/x-nikon-nef: { schema: { type: string, format: binary } }
            application/octet-stream: { schema: { type: string, format: binary } }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/camera/focus:
    post:
      tags: [camera]
      summary: Nudge the lens focus motor
      operationId: postCameraFocus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusNudgeRequest"
      responses:
        "204":
          description: Nudge issued.
        "503":
          $ref: "#/components/responses/CameraUnavailable"
...
  /api/camera/preview/stream:
    get:
      tags: [camera]
      summary: Continuous MJPEG preview stream (browser-friendly)
      description: |
        Returns `multipart/x-mixed-replace; boundary=frame`. Suitable for an
        `<img>` tag. Capped server-side to ≤ 5 fps on Pi Zero 2 W to keep CPU
        below 30 % during preview. Closing the connection stops streaming.
      operationId: getCameraPreviewStream
      responses:
        "200":
          description: MJPEG stream.
          content:
            multipart/x-mixed-replace:
              schema: { type: string, format: binary }
        "503":
          $ref: "#/components/responses/CameraUnavailable"
...
components:
  responses:
    ValidationError:
      description: Request body or query failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    CameraUnavailable:
      description: Camera is disconnected or in error state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code (snake_case).
          examples:
            - validation_failed
            - job_already_running
            - calibration_required
            - camera_disconnected
            - tilt_limit_exceeded
            - protocol_version_mismatch
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    CameraSettings:
      type: object
      description: Map of gphoto2 widget path → current value/type/choices.
      additionalProperties:
        type: object
        required: [type, current]
        properties:
          type:
            type: string
            enum: [TEXT, RANGE, TOGGLE, RADIO, MENU, DATE, BUTTON, UNKNOWN]
          current:
            description: Current value (string|number|bool|null).
            oneOf:
              - { type: string }
              - { type: number }
              - { type: boolean }
              - { type: "null" }
          choices:
            oneOf:
              - { type: array, items: { type: string } }
              - { type: "null" }
          range:
            type: object
            nullable: true
            properties:
              min: { type: number }
              max: { type: number }
              step: { type: number }
    CameraSettingsUpdate:
      type: object
      required: [settings]
      properties:
        settings:
          type: object
          additionalProperties:
            oneOf:
              - { type: string }
              - { type: number }
              - { type: boolean }
        step_policy:
          type: string
          enum: [strict, snap]
          default: strict
    CaptureRequest:
      type: object
      properties:
        autofocus:
          type: boolean
          default: false
          description: When `false`, skip the AF motor (matches `capture_image_no_af`).
        save_to_session:
          oneOf:
            - { type: string, format: uuid, description: "Existing session_id" }
            - { type: "null" }
          description: Optional — store the test capture as an asset on a session.
    FocusNudgeRequest:
      type: object
      required: [step_size]
      properties:
        step_size:
          type: integer
          minimum: -3
          maximum: 3
          description: |
            Incremental focus step. Positive for 'near', negative for 'far'.
            Scale is usually 1=small, 2=medium, 3=large.
    CaptureResult:
      type: object
      required: [capture_id, content_type, captured_at, size_bytes]
      properties:
        capture_id: { type: string, format: uuid }
        content_type: { type: string }
        captured_at: { type: string, format: date-time }
        size_bytes: { type: integer }
        download_url:
          type: string
          format: uri-reference
          description: Path to fetch the binary (`/api/camera/captures/{id}`).
...